Move config for Interval Reference DAG. (#436)
* Update clinvar download to run in dev environment

* select fields we care abour

* add back old one

* preserve old

* single line

* update versioning

* chmod

* ws

* a few changes with fresh eyes

* cleanup

* Update constants.py

* remove os

* Update clinvar.py

* Update write_clinvar_ht.py

* Move config and constants

* ws

* pr changes

* remember delimiter :facepalm:

* cleanup

* better

* annotate

* hailify

* I guess this one should be camelcase

* remove print

* getting to a stopping point

* more

* Annotate the tables

* reforamt

* remove clinvar

* function

* don't move version

* unused
